requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
get==1.0.3
idna==2.7
lxml==4.2.5
post==1.0.2
public==2.0.1
pycrypto==2.6.1
query-string==1.0.2
request==1.0.2
requests==2.20.0
selenium==3.14.1
urllib3==1.24
